Surrogate Model Asisted Synchronized Hyperband Optimizer

Perform Surrogate Model Assisted Hyperband Optimization.

Given a population size mu, a fraction of surviving individuals survival_fraction, a number of generations n and a fidelity progression F1, F1, ..., Fn, the algorithm works as follows:

Sample an initial design of size mu at fidelity F1

Kill individuals that are not in the top survival_fraction part of individuals, by performance

Generate new individuals using random sampling and optionally a Filtor, until mu alive individuals are present again.

Evaluate all alive individuals at fidelity F{generation}

Jump to 2., until termination, possibly because n generations are reached.

The number of generations n is determined from the OptimInstance's Terminator object, see Terminating below.

The fidelity progression uses a specially designated "budget" parameter of the OptimInstance, which must have a "budget" tag. The lower limit of the budget parameter is used as F0. The upper limit is used as Fn. The budget is evaluated at equally spaced values in generations 0..n, so F1 - F0 = F2 - F1 etc. In many cases it is desirable to have a multiplicative progression of "difficulty" of the problem. In this case, it is recommended to use a budget parameter with exponential "trafo", or one with logscale = TRUE (see example).

A Filtor can be used for filtering, see the respective class's documentation for details on algorithms. The FiltorSurrogateProgressive can be used for progressive surrogate model filtering. FiltorMaybe can be used for random interleaving.

Fidelity Steps

The number of fidelity steps can be determined through the fidelity_steps configuration parameter, or can be determined when a TerminatorGenerations is used to determine the number of fidelity refinements that are being performed. For this, the OptimInstance being optimized must contain a TerminatorGenerations, either directly (inst$terminator), or indirectly through a bbotk::TerminatorCombo with $any set to TRUE (recursive TerminatorCombo may also be used).

Configuration Parameters

OptimizerSmashy's configuration parameters are the hyperparameters of the Filtor given to the filtor construction argument, as well as:

mu :: integer(1)
Population size: Number of individuals that are sampled in the beginning, and which are re-evaluated in each fidelity step. Initialized to 2.

survival_fraction :: numeric(1)
Fraction of the population that survives at each fidelity step. The number of newly sampled individuals is (1 - survival_fraction) * mu. Initialized to 0.5.

sampling :: function
Function that generates the initial population, as well as new individuals to be filtered from, as a Design object. The function must have arguments param_set and n and function like paradox::generate_design_random or paradox::generate_design_lhs. This is equivalent to the initializer parameter of mies_init_population(), see there for more information. Initialized to generate_design_random().

fidelity_steps :: integer(1)
Number of fidelity steps, must be at least 1.

synchronize_batches :: logical(1)
Whether to use synchronized batches, as opposed to "Hyperband"-like successive halfing. Initialized to TRUE.

filter_with_max_budget :: logical(1)
Whether to perform filtering with the maximum fidelity value found in the archive, as opposed the current budget_survivors. This has only an effect when fidelity_steps is greater than 1 and some evaluations are done when the archive already contains evaluations with greater fidelity. Initialized to FALSE.

See also

Super class

bbotk::Optimizer -> OptimizerSmashy

Active bindings

filtor

(Filtor)
Filtering algorithm used.

selector

(Selector)
Survival selector used between fidelity steps.

param_set

(ParamSet)
Configuration parameters of the optimization algorithm.

Methods

Public methods

OptimizerSmashy$new()

OptimizerSmashy$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Initialize the 'OptimizerSmashy' object.

Usage

OptimizerSmashy$new(filtor = FiltorProxy$new(), selector = SelectorProxy$new())

Arguments

filtor

(Filtor)
Filtor for the filtering algorithm. Default is FiltorProxy, which exposes the operation as a configuration parameter of the optimizer itself.
The $filtor field will reflect this value.

selector

(Selector)
Selector that chooses which individuals survive to the next higher fidelity step. Note this is potentially different from the selection done in filtor (although the common case would be to use the same Selector in both). The $selector field will reflect this value.

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSmashy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerSmashy

Methods

Public methods

TunerSmashy$new()

TunerSmashy$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Initialize the TunerSmashy object.

Usage

TunerSmashy$new(filtor = FiltorProxy$new(), selector = SelectorProxy$new())

Arguments

surrogate_learner

(mlr3::LearnerRegr | NULL)

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerSmashy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
lgr::threshold("warn")

#####
# Optimizing a Function
#####

library("bbotk")

# Define the objective to optimize
# The 'budget' here simulates averaging 'b' samples from a noisy function
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    z <- z + rnorm(1, sd = 1 / sqrt(xs$b))
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4), b = p_int(1)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

search_space = objective$domain$search_space(list(
  x = to_tune(),
  y = to_tune(),
  b = to_tune(p_int(1, 2^10, logscale = TRUE, tags = "budget"))
))

# Get a new OptimInstance. Here we determine that the optimizatoin goes
# for 10 generations.
oi <- OptimInstanceSingleCrit$new(objective,
  search_space = search_space,
  terminator = trm("gens", generations = 10)
)

library("mlr3learners")
# use the 'regr.ranger' as surrogate.
# The following settings have 30 individuals in a batch, the 20 best
# of which survive, while 10 are sampled new.
# For this, 100 individuals are sampled randomly, and the top 10, according
# to the surrogate model, are used.
smashy_opt <- opt("smashy", ftr("surprog",
    surrogate_learner = mlr3::lrn("regr.ranger"),
    filter.pool_factor = 10), fidelity_steps = 10,
  mu = 30, survival_fraction = 2/3
)
# smashy_opt$optimize performs Smashy optimization and returns the optimum
smashy_opt$optimize(oi)
#>           x        y        b  x_domain     Obj
#> 1: 2.301525 1.669191 2.310816 <list[3]> 2.32829

#####
# Optimizing a Machine Learning Method
#####

# Note that this is a short example, aiming at clarity and short runtime.
# The settings are not optimal for hyperparameter tuning.

library("mlr3")
library("mlr3learners")
library("mlr3tuning")

# The Learner to optimize
learner = lrn("classif.xgboost")

# The hyperparameters to optimize
learner$param_set$values[c("eta", "booster")] = list(to_tune())
learner$param_set$values$nrounds = to_tune(p_int(1, 4, tags = "budget", logscale = TRUE))

# Get a TuningInstance
ti = TuningInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.acc"),
  terminator = trm("gens", generations = 3)
)

# use ftr("maybe") for random interleaving: only 50% of proposed points are filtered.
smashy_tune <- tnr("smashy", ftr("maybe", p = 0.5, filtor = ftr("surprog",
    surrogate_learner = lrn("regr.ranger"),
    filter.pool_factor = 10)), fidelity_steps = 10,
  mu = 20, survival_fraction = 0.5
)
# smashy_tune$optimize performs Smashy optimization and returns the optimum
smashy_tune$optimize(ti)
#>      nrounds       eta booster learner_param_vals  x_domain classif.acc
#> 1: 0.1788264 0.6554934  gbtree          <list[5]> <list[3]>        0.98

# }