Density Estimation Prediction Object

Predictions returned by LearnerDensity and evaluated by MeasureDensity.

See also

Super class

mlr3::Prediction -> PredictionDensity

Active bindings

prob

(numeric)
Predicted density.

prob

(numeric)
Predicted log density.

Methods

Public methods

PredictionDensity$new()

Inherited methods

mlr3::Prediction$filter()

mlr3::Prediction$format()

mlr3::Prediction$help()

mlr3::Prediction$print()

mlr3::Prediction$score()

Method new()

Initialize the PredictionDensity object. One of prob or logprob must be given.

Usage

PredictionDensity$new(
  task = NULL,
  row_ids = mlr3::assert_task(task)$row_ids,
  prob = exp(logprob),
  logprob = log(prob),
  check = TRUE
)

Arguments

task

(TaskDensity)
TaskDensity for which prediction was made, used to extract defaults for row_ids.

row_ids

(integer)
Row ids of the predicted observations.

prob

(numeric)
Predicted density

logprob

(numeric)
Predicted log density.

check

(logical(1))
Whether to perform argument checks and predict type conversions.