Put Budget-Parameter on Logarithmic Scale

Set the search space component that is tagged as "budget" to log-scale.

Smashy operates on a log-scale internally, so the budget of the search space must also be on a log-scale, and should be transformed (through exponentiation) by the $trafo() function of the ParamSet.

budget_to_logscale() does this automatically: It takes a ParamSet with unscaled budget (e.g. number of deep learning epochs, or timeout), and returns a ParamSet where this configuration parameter is replaced by another configuration parameter with lower and upper bounds replaced by the log() of the original bounds. Furthermore, the $trafo() is adjusted to include exponentiation of the budget. Another $trafo() function may be present, the exponentiation of the budget parameter is then followed by a call to the original $trafo().

budget_to_logscale(search_space)

Arguments

search_space

(ParamSet)
Search space to modify. The input is not modified in-place and is cloned instead.

Value

ParamSet: The search space with the "budget"-tagged component on a log-scale.

See also

Examples

search_space = ps(
  b = p_int(1, 100, tags = "budget")
)

ss_t = budget_to_logscale(search_space)

library("data.table")
d = Design$new(ss_t, data.table(b = log(c(1, 10, 100))), FALSE)

d
#> <Design> with 3 rows:
#>           b
#> 1: 0.000000
#> 2: 2.302585
#> 3: 4.605170

d$transpose()
#> [[1]]
#> [[1]]$b
#> [1] 1
#> 
#> 
#> [[2]]
#> [[2]]$b
#> [1] 10
#> 
#> 
#> [[3]]
#> [[3]]$b
#> [1] 100
#> 
#>