Configure Full Smashy Optimizer.

Create a OptimizerSmashy (or TunerSmashy) object with extended functionality from basic building blocks. While basic Smashy-optimizers can be created with opt("smashy") and tnr("smashy"), setting them up for functionality such as BOHB-like surrogate model filtering or change of filter configuration over the course of optimization can be difficult, hence this helper function. configure_smashy() should be used when the extended functionality is desired.

configure_smashy(
  search_space,
  budget_log_step,
  mu,
  survival_fraction = 1/3,
  sample = "bohb",
  batch_method = "smashy",
  filter_algorithm = "tournament",
  surrogate_learner = "knn1",
  filter_with_max_budget = TRUE,
  filter_factor_first = 100,
  filter_factor_last = filter_factor_first,
  filter_select_per_tournament = 1,
  random_interleave_fraction = 1/3,
  filter_factor_first.end = filter_factor_first,
  filter_factor_last.end = filter_factor_last,
  filter_select_per_tournament.end = filter_select_per_tournament,
  random_interleave_fraction.end = random_interleave_fraction,
  random_interleave_random = TRUE,
  type = "Optimizer"
)

Arguments

search_space

(ParamSet)
search space for which OptimizerSmashy should be configured, has one parameter tagged 'budget'. Must be the same search space as of the OptimInstanceSingleCrit being optimized with the resulting optimizer.
The budget-parameter must be in log-scale; in most cases this means that budget_to_logscale() must be called on the search space. Note that this should be done before OptimInstanceSingleCrit is created with it, see examples.

budget_log_step

(numeric(1))
log() of budget fidelity steps to make. E.g. log(2) for doubling. Must always be on log scale.

mu

(integer(1))
Population size, must be at least 2.

survival_fraction

(numeric(1))
Fraction of individuals that survive each round. Must be between 0 and 1. If mu is very large, it is decreased so that round(mu * survival_fraction) is < survival_fraction. Defaults to 1 / 3.

sample

(character(1))
Sampling mechanism: Must be either "random" (sample uniformly at random), or "bohb" (use BOHB sampling from kernel density estimate of well performing points). Default "bohb".

batch_method

(character(1))
Fidelity schedule, one of "smashy" or "hb". "smashy" uses synchronized batches, "hb" does scheduling similar to Hyperband, but generalized for possibly disagreeing values of budget_log_step and survival_fraction. For true Hyperband (up to rounding), set budget_log_step to log(survival_fraction) and mu to survival_fraction ^ -fidelity_steps (where fidelity_steps counts from 0 for "no step, only highest budget is evaluated"). Defaults to "smashy".

filter_algorithm

(character(1))
One of "tournament" (using FiltorSurrogateTournament) or "progressive" (using FiltorSurrogateProgressive). Defaults to "tournament".

surrogate_learner

(character(1))
Surrogate learner to use for filtering. One of "knn1", "knn7" (KNN with k = 1 or k = 7, respectively), "ranger" (random forest) or "bohblrn" (emulating the TPE used in the BOHB paper). Defaults to "knn1".

filter_with_max_budget

(logical(1))
Whether to use the surrogate learner to predict performance at the maximum budget value that has been evaluated thus far, or at the budget value being currently evaluated. Behavour of these only disagrees once at least one full-budget configuration has been evaluated. Default TRUE.

filter_factor_first

(numeric(1))
When filtering the first configuration using the FiltorSurrogateTournament or FiltorSurrogateProgressive, how many random samples to consider. Corresponds to filter.tournament_size/filter.per_tournament in FiltorSurrogateTournament and filter.pool_factor in FiltorSurrogateProgressive. Default 100.
Note the Simulated Annealing section about filter_factor_first.end.

filter_factor_last

(numeric(1))
When filtering the last configuration using the FiltorSurrogateTournament or FiltorSurrogateProgressive, how many random samples to consider. Corresponds to filter.tournament_size_last/filter.per_tournament in FiltorSurrogateTournament and filter.pool_factor_last in FiltorSurrogateProgressive. Default to the value given to filter_factor_first. Note the Simulated Annealing section about filter_factor_last.end.

filter_select_per_tournament

(integer(1))
Only when filter_algorithm is "tournament": How many configurations to draw per tournament. Corresponds to filter.per_tournament of FiltorSurrogateTournament. Must be at least 1. Defaults to 1. Note the Simulated Annealing section about filter_select_per_tournament.end.

random_interleave_fraction

(numeric(1))
How many configurations to not filter using the surrogate in each generation and to instead draw randomly using the sample method. Must be between 0 and 1, where a value of 1 means no surrogate-based filtering is done at all, all values are drawn randomly. Defaults to 1 / 3.
Note the Simulated Annealing section about random_interleave_fraction.end.

filter_factor_first.end

(numeric(1))
Corresponds to filter_factor_first, see the Simulated Annealing section.

filter_factor_last.end

(numeric(1))
Corresponds to filter_factor_last, see the Simulated Annealing section.

filter_select_per_tournament.end

(integer(1))
Corresponds to filter_select_per_tournament, see the Simulated Annealing section.

random_interleave_fraction.end

(numeric(1))
Corresponds to random_interleave_fraction, see the Simulated Annealing section.

random_interleave_random

(logical(1))
Whether during each sampling, a fixed fraction of random_interleave_fraction individuals is drawn randomly (FALSE), or the number of individuals to draw randomly is itself random (TRUE). In the latter case, each new configuration has an independent chance of being drawn randomly with probability random_interleave_fraction. Default TRUE.

type

(character(1))
One of "Optimizer" or "Tuner". What class to return: An Optimizer for optimizing OptimInstanceSingleCrit objects, or a Tuner for tuning mlr3::Learners. Defaults to "Optimizer".

Value

OptimizerSmashy or TunerSmashy, depending on the type argument: The configured Smashy optimizer.

Simulated Annealing

For the function arguments filter_factor_first, filter_factor_last, select_per_tournament, and random_interleave_fraction, it is possible to also give corresponding *.end arguments. When the corresponding *.end argument differs from the given value, then the effective value changes over the course of optimization, starting at the given value and progressing geometrically (for filter_factor_first, filter_factor_last, and select_per_tournament) or linearly (for random_interleave_fraction) towards the *.end-value at the end.

See also

Examples

# Define the objective to optimize
# The 'budget' here simulates averaging 'b' samples from a noisy function
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    z <- z + rnorm(1, sd = 1 / sqrt(xs$b))
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4), b = p_int(1)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

search_space_proto = objective$domain$search_space(list(
  x = to_tune(),
  y = to_tune(),
  b = to_tune(p_int(1, 2^10, tags = "budget"))
))

search_space = budget_to_logscale(search_space_proto)

# Get a new OptimInstance. Here we determine that the optimizatoin goes
# for 10 full budget evaluations (10 * 2^100)
oi <- OptimInstanceSingleCrit$new(objective,
  search_space = search_space,
  terminator = trm("budget", budget = 10 * 2^10)
)

# smashy is designed with parallel evaluation in mind. It is
# recommended to run
# > future::plan("multisession") (windows, or in RStudio)
# or
# > future::plan("multicore") (Linux without RStudio)
# and then set `mu` to a multiple of the number of CPU cores.

smashy = configure_smashy(search_space, budget_log_step = log(4), mu = 6)

smashy$optimize(oi)
#>           x        y b  x_domain      Obj
#> 1: 1.769893 1.975496 0 <list[3]> 3.340931