Simulated Binary Crossover Recombinator

Numeric Values between two individuals are recombined via component-wise independent simulated binary crossover. See Deb (1995) for more details.

Configuration Parameters

p :: numeric(1)
Component-wise probability with which to crossover values. Initialized to 0.5.

n :: double
Non-negative distribution index of the polynomial distribution for each component. Generally spoken, the higher n, the higher the probability of creating near parent values. This may either be a scalar in which case it is applied to all input components, or a vector, in which case it must have the length of the input components and applies to components in order in which they appear in the priming ParamSet. Initialized to 1.

Supported Operand Types

Supported Param classes are: ParamDbl

Dictionary

This Recombinator can be created with the short access form rec() (recs() to get a list), or through the the dictionary dict_recombinators in the following way:

# preferred:
rec("sbx")
recs("sbx")  # takes vector IDs, returns list of Recombinators

# long form:
dict_recombinators$get("sbx")

References

Deb, Kalyanmoy, Agrawal, Bhushan R, others (1995). “Simulated binary crossover for continuous search space.” Complex systems, 9(2), 115--148.

See also

Super classes

smashy::MiesOperator -> smashy::Recombinator -> RecombinatorSimulatedBinaryCrossover

Methods

Public methods

RecombinatorSimulatedBinaryCrossover$new()

RecombinatorSimulatedBinaryCrossover$clone()

Inherited methods

smashy::MiesOperator$operate()

smashy::MiesOperator$prime()

smashy::MiesOperator$print()

smashy::MiesOperator$repr()

Method new()

Initialize the RecombinatorSimulatedBinaryCrossover object.

Usage

RecombinatorSimulatedBinaryCrossover$new(keep_complement = TRUE)

Method clone()

The objects of this class are cloneable with this method.

Usage

RecombinatorSimulatedBinaryCrossover$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

set.seed(1)
rsbx = rec("sbx")
p = ps(x = p_dbl(-5, 5), y = p_dbl(-5, 5), z = p_dbl(-5, 5))
data = data.frame(x = 0:5, y = 0:5, z = 0:5)

rsbx$prime(p)
rsbx$operate(data)
#>            x          y        z
#> 1 1.03873632  1.9296682 0.000000
#> 2 0.06954315 -0.5894168 1.000000
#> 3 2.79413060  2.9938346 2.000000
#> 4 2.17944045  1.7657679 3.000000
#> 5 3.75236703  4.0000000 4.134726
#> 6 4.98340077  5.0000000 4.810690

rsbx$param_set$values$n = c(0.5, 1, 10)
rsbx$param_set$values$p = 1
rsbx$operate(data)
#>            x          y           z
#> 1 -0.6560039 0.80419411  0.93495586
#> 2  0.8821153 0.00426727 -0.04203252
#> 3  3.0984128 1.98508719  1.97077831
#> 4  1.5337509 3.24299718  3.02262829
#> 5  4.0440152 4.86008133  3.91392295
#> 6  4.6956626 4.10261250  4.94728532