Evaluate Proposed Configurations Generated in a MIES Iteration

Calls $eval_batch of a given OptimInstance on a set of configurations as part of a MIES operation. The dob extra-info in the archive is also set properly to indicate a progressed generation.

This function can be used directly, but it is easier to use it within the OptimizerMies Optimizer if standard GA operation is desired.

"Rolling-tide" multifidelity is supported as described in OptimizerMies or more thoroughly in vignette("mies-multifid"). For this, an extra component named after budget_id is appended to each individual, chosen from the fidelity_schedule depending on the value of survivor_budget. Both fidelity_schedule and budget_id should be the same values as given to the other mies_* functions.

mies_evaluate_offspring(
  inst,
  offspring,
  fidelity_schedule = NULL,
  budget_id = NULL,
  survivor_budget = FALSE,
  step_fidelity = FALSE,
  monotonic = TRUE
)

Arguments

inst

(OptimInstance)
Optimization instance to evaluate.

offspring

(data.frame)
Proposed configurations to be evaluated, must have columns named after the inst's search space, minus budget_id if not NULL.

fidelity_schedule

(data.frame | NULL)
data.frame with three columns "generation", "budget_new", "budget_survivors", in that order. "budget_new" and "budget_survivors" are atomic columns assigned to the budget_id component of offspring; which one is chosen depends on survivor_budget. "generation" is an integer valued column, indicating the first generation at which a row is valid. At least one row with generation == 1 must be present.
This value may be NULL if no multi-fidelity optimization is performed (the default).
Note that the multifidelity functionality is experimental and the UI may change in the future.

budget_id

(character(1) | NULL)
Budget component that is set to the value found in fidelity_schedule when doing multi-fidelity optimization. Should be NULL when and only when fidelity_schedule is NULL.

survivor_budget

(logical(1))
When doing multi-fidelity optimization, determines which column of fidelity_schedule to use to determine the budget component value.
Note that the multifidelity functionality is experimental and the UI may change in the future.

step_fidelity

(logical(1))
Whether to also evaluate alive individuals from previous generations that have a budget value below the current budget_survivors. Default FALSE. Ignored when budget_id and fidelity_schedule are NULL.
Note that the multifidelity functionality is experimental and the UI may change in the future.

monotonic

(logical(1))
When step_fidelity is TRUE, then this indicates whether individuals should only ever be re-evaluated when fidelity would be increased. Ignored when step_fidelity is FALSE or when budget_id and fidelity_schedule are NULL.
Note that the multifidelity functionality is experimental and the UI may change in the future.

Value

invisible

data.table: the performance values returned when evaluating the offspring values through eval_batch.

See also

Examples

library("bbotk")
lgr::threshold("warn")

# Define the objective to optimize
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

# Get a new OptimInstance
oi <- OptimInstanceSingleCrit$new(objective,
  terminator = trm("evals", n_evals = 100)
)

mies_init_population(inst = oi, mu = 3)
# Initial state:
oi$archive
#> <Archive>
#>        x    y dob eol     Obj              timestamp batch_nr
#> 1:  2.66  1.9   1  NA 1.3e+00 2022-08-10 05:17:42.50        1
#> 2:  3.61 -1.2   1  NA 4.4e-06 2022-08-10 05:17:42.50        1
#> 3: -0.73 -0.4   1  NA 5.0e-01 2022-08-10 05:17:42.50        1

# 'offspring' is just a data.frame of values to evaluate.
# In general it should be created using 'mies_generate_offspring()'.
offspring = data.frame(x = 1:2, y = 2:1)

mies_evaluate_offspring(oi, offspring = offspring)

# This evaluated the given points and assigned them 'dob' 2.
oi$archive
#> <Archive>
#>        x    y dob eol     Obj              timestamp batch_nr
#> 1:  2.66  1.9   1  NA 1.3e+00 2022-08-10 05:17:42.50        1
#> 2:  3.61 -1.2   1  NA 4.4e-06 2022-08-10 05:17:42.50        1
#> 3: -0.73 -0.4   1  NA 5.0e-01 2022-08-10 05:17:42.50        1
#> 4:  1.00  2.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2
#> 5:  2.00  1.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2

# Note that at this point one would ordinarily call a 'mies_survival_*()'
# function.

###
# Advanced demo, making use of additional components and fidelity_schedule
# Note that the multifidelity functionality is experimental and the UI may
# change in the future.
##

# declare 'y' the budget parameter. It does not occur in 'offspring' any more.
budget_id = "y"
# but: offspring may contain any other value that is appended to 'oi'. They
# are ignored by the objective.
offspring = data.frame(x = 0:1, z = 3)

fidelity_schedule = data.frame(
  generation = c(1, 5),
  budget_new = c(1, 2),
  budget_survivors = c(3, 4)
)
# The fidelity schedule:
fidelity_schedule
#>   generation budget_new budget_survivors
#> 1          1          1                3
#> 2          5          2                4

# When given a fidelity_schedule and budget_id, the 'budget_new' budget
# is used. Here it is '1', because the geeneration number 3 is being evaluated.
mies_evaluate_offspring(oi, offspring = offspring,
  fidelity_schedule = fidelity_schedule, budget_id = budget_id)

# This now has the additional column 'z'. Values of y for the new evaluations
# are 1.
oi$archive
#> <Archive>
#>        x    y dob eol     Obj              timestamp batch_nr  z
#> 1:  2.66  1.9   1  NA 1.3e+00 2022-08-10 05:17:42.50        1 NA
#> 2:  3.61 -1.2   1  NA 4.4e-06 2022-08-10 05:17:42.50        1 NA
#> 3: -0.73 -0.4   1  NA 5.0e-01 2022-08-10 05:17:42.50        1 NA
#> 4:  1.00  2.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2 NA
#> 5:  2.00  1.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2 NA
#> 6:  0.00  1.0   3  NA 3.8e-01 2022-08-10 05:17:42.54        3  3
#> 7:  1.00  1.0   3  NA 4.1e-01 2022-08-10 05:17:42.54        3  3

# setting 'survivor_budget = TRUE' uses the 'budget_survivors' schedule column.
mies_evaluate_offspring(oi, offspring = offspring,
  fidelity_schedule = fidelity_schedule, budget_id = budget_id,
  survivor_budget = TRUE)

# the new row has y set to 2.
oi$archive
#> <Archive>
#>        x    y dob eol     Obj              timestamp batch_nr  z
#> 1:  2.66  1.9   1  NA 1.3e+00 2022-08-10 05:17:42.50        1 NA
#> 2:  3.61 -1.2   1  NA 4.4e-06 2022-08-10 05:17:42.50        1 NA
#> 3: -0.73 -0.4   1  NA 5.0e-01 2022-08-10 05:17:42.50        1 NA
#> 4:  1.00  2.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2 NA
#> 5:  2.00  1.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2 NA
#> 6:  0.00  1.0   3  NA 3.8e-01 2022-08-10 05:17:42.54        3  3
#> 7:  1.00  1.0   3  NA 4.1e-01 2022-08-10 05:17:42.54        3  3
#> 8:  0.00  3.0   4  NA 1.4e-02 2022-08-10 05:17:42.56        4  3
#> 9:  1.00  3.0   4  NA 2.7e-01 2022-08-10 05:17:42.56        4  3

# When calling 'mies_evaluate_offspring()' with the same 'fidelity_schedule'
# now, then the generation 5 is evaluated. Therefore the next row of
# the schedule is used, and 'y' is therefore set to 2 (or 4, ir survivor budget
# is used).

mies_evaluate_offspring(oi, offspring = offspring,
  fidelity_schedule = fidelity_schedule, budget_id = budget_id)

oi$archive
#> <Archive>
#>         x    y dob eol     Obj              timestamp batch_nr  z
#>  1:  2.66  1.9   1  NA 1.3e+00 2022-08-10 05:17:42.50        1 NA
#>  2:  3.61 -1.2   1  NA 4.4e-06 2022-08-10 05:17:42.50        1 NA
#>  3: -0.73 -0.4   1  NA 5.0e-01 2022-08-10 05:17:42.50        1 NA
#>  4:  1.00  2.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2 NA
#>  5:  2.00  1.0   2  NA 7.4e-01 2022-08-10 05:17:42.52        2 NA
#>  6:  0.00  1.0   3  NA 3.8e-01 2022-08-10 05:17:42.54        3  3
#>  7:  1.00  1.0   3  NA 4.1e-01 2022-08-10 05:17:42.54        3  3
#>  8:  0.00  3.0   4  NA 1.4e-02 2022-08-10 05:17:42.56        4  3
#>  9:  1.00  3.0   4  NA 2.7e-01 2022-08-10 05:17:42.56        4  3
#> 10:  0.00  2.0   5  NA 5.5e-02 2022-08-10 05:17:42.58        5  3
#> 11:  1.00  2.0   5  NA 7.4e-01 2022-08-10 05:17:42.58        5  3
# TODO example with step_fidelity = TRUE