Re-Evaluate Configurations with Higher Fidelity

As part of the "rolling-tide" multifidelity-setup, do reevaluation of configurations with higher fidelity that have survived lower-fidelity selection.

The fidelity_schedule specifies what the fidelity (i.e. value of the "budget" component) of evaluations done or compared at generation i should be. The generation_lookahead parameter has a subtle influence. If it is TRUE, then, at the end of generation i, reevaluation is done with the "budget_survivors" fidelity of generation i + 1, otherwise the entry of generation i is used. This gives the "budget_survivors" column a slightly different meaning in the two cases: either it is the budget, with which survivors "entering" generation i have been evaluated, or it is the budget with which survivors "coming out of" that generation were evaluated. The evaluations are done as part of the current generation in any case (see examples).

This function should only be called when doing rolling-tide multifidelity, and should not be part of the MIES cycle otherwise.

Note that the multifidelity functionality is experimental and the UI may change in the future.

mies_step_fidelity(
  inst,
  fidelity_schedule,
  budget_id,
  generation_lookahead = TRUE,
  current_gen_only = FALSE,
  monotonic = TRUE,
  additional_components = NULL
)

Arguments

inst

(OptimInstance)
Optimization instance to evaluate.

fidelity_schedule

(data.frame)
data.frame with three columns "generation", "budget_new", "budget_survivors", in that order. "budget_new" and "budget_survivors" are atomic columns. The value of "budget_survivors" of the appropriate row is assigned to the budget_id component of offspring. "generation" is an integer valued column, indicating the first generation at which a row is valid. At least one row with generation == 1 must be present.

budget_id

(character(1))
Budget component that is set to the value found in fidelity_schedule.

generation_lookahead

(logical(1))
Whether to re-evaluate alive individuals with the "budget_survivors" entry of the current generation (FALSE), or of the next generation (TRUE).

current_gen_only

(logical(1))
Whether to only re-evaluate survivors individuals generated in the latest generation (TRUE), or re-evaluate all currently alive individuals (FALSE). In any case, only individuals that were not already evaluated with the chosen fidelity are evaluated, so this will usually only affect individuals from before the latest generation if the "budget_survivors" changed between generations (i.e. when a different row from fidelity_schedule is used).

monotonic

(logical(1))
Whether to only re-evaluate configurations for which the fidelity would increase. Default TRUE.

additional_components

(ParamSet | NULL)
Additional components to optimize over, not included in search_space, but possibly used for self-adaption. This must be the ParamSet of mies_init_population()'s additional_component_sampler argument.

Value

invisible

data.table: the performance values returned when evaluating the offspring values through eval_batch.

See also

Examples

library("bbotk")
lgr::threshold("warn")

# Define the objective to optimize
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

# Get a new OptimInstance
oi <- OptimInstanceSingleCrit$new(objective,
  terminator = trm("evals", n_evals = 100)
)

budget_id = "y"

fidelity_schedule = data.frame(
  generation = c(1, 2),
  budget_new = c(1, 3),
  budget_survivors = c(2, 4)
)

# The initial population gets evaluated with 'budget_survivors', so y = 2.
mies_init_population(oi, mu = 2, fidelity_schedule = fidelity_schedule,
  budget_id = budget_id)

oi$archive
#> <Archive>
#>        x y dob eol   Obj              timestamp batch_nr
#> 1: -0.40 2   1  NA 0.022 2022-08-10 05:17:46.56        1
#> 2:  0.32 2   1  NA 0.134 2022-08-10 05:17:46.56        1

# By default, mies_step_fidelity() will re-evaluate all survivors with
# budget_survivors of the *next* generation. This is the reason why it
# usually does not make sense to have 'budget_survivors' change between
# generations 1 and 2!
mies_step_fidelity(oi, fidelity_schedule = fidelity_schedule,
  budget_id = budget_id)

# The first two evaluations have 'eol = 1' (died in generation 1) and were
# re-evaluated with y = 4.
oi$archive
#> <Archive>
#>        x y dob eol     Obj              timestamp batch_nr
#> 1: -0.40 2   1   1 0.02205 2022-08-10 05:17:46.56        1
#> 2:  0.32 2   1   1 0.13418 2022-08-10 05:17:46.56        1
#> 3: -0.40 4   1  NA 0.00012 2022-08-10 05:17:46.57        2
#> 4:  0.32 4   1  NA 0.00215 2022-08-10 05:17:46.57        2

# Using 'generation_lookahead = FALSE' will use the 'budget_survivors' of the
# current genereation, where the current generation is always the max() of the
# 'dob' column in the archive.
# Calling 'mies_step_fidelity()' multiple times within one generation barely
# ever makes sense and is done here for demonstration.
# The following call will have *no* effect: The 'budget_survivors' of the
# current generation (generation 1) is 2, but the alive individuals were
# evaluated with budget 4.
mies_step_fidelity(oi, fidelity_schedule = fidelity_schedule,
  budget_id = budget_id, generation_lookahead = FALSE)

oi$archive
#> <Archive>
#>        x y dob eol     Obj              timestamp batch_nr
#> 1: -0.40 2   1   1 0.02205 2022-08-10 05:17:46.56        1
#> 2:  0.32 2   1   1 0.13418 2022-08-10 05:17:46.56        1
#> 3: -0.40 4   1  NA 0.00012 2022-08-10 05:17:46.57        2
#> 4:  0.32 4   1  NA 0.00215 2022-08-10 05:17:46.57        2

# To also re-evaluate individuals with *lower* budget, use 'monotonic = FALSE':
mies_step_fidelity(oi, fidelity_schedule = fidelity_schedule,
  budget_id = budget_id, generation_lookahead = FALSE, monotonic = FALSE)

oi$archive
#> <Archive>
#>        x y dob eol     Obj              timestamp batch_nr
#> 1: -0.40 2   1   1 0.02205 2022-08-10 05:17:46.56        1
#> 2:  0.32 2   1   1 0.13418 2022-08-10 05:17:46.56        1
#> 3: -0.40 4   1   1 0.00012 2022-08-10 05:17:46.57        2
#> 4:  0.32 4   1   1 0.00215 2022-08-10 05:17:46.57        2
#> 5: -0.40 2   1  NA 0.02205 2022-08-10 05:17:46.60        3
#> 6:  0.32 2   1  NA 0.13418 2022-08-10 05:17:46.60        3