Featureless Density Estimation Learner

Simple LearnerDensity fallback learner which gives all predictions the density .Machine$double.xmin (typically 2.225074e-308).

# TODO: could make this a hyperparameter.

See also

Super classes

mlr3::Learner -> smashy::LearnerDensity -> LearnerDensityFeatureless

Methods

Public methods

LearnerDensityFeatureless$new()

LearnerDensityFeatureless$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

smashy::LearnerDensity$sample()

Method new()

Initialize the LearnerDensityFeatureless object.

Usage

LearnerDensityFeatureless$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerDensityFeatureless$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.