Kernel Density Estimator using the "np" Package

Perform kernel density estimation using the np package.

Hyperparameters

bwmethod :: character(1)
Bandwidth selection method. One of "cv.ml", "cv.ls", "normal-reference", "normal-reference-numeric". Default "cv.ml". "normal-reference-numeric" emulates (buggy, see https://github.com/statsmodels/statsmodels/issues/3790) behaviour of python statsmodels: non-numeric columns are treated as numeric for bandwidth estimation.

bwtype :: character(1)
Continuous variable bandwidth type, one of "fixed" (default), "generalized_nn", "adaptive_nn".

ckertype :: character(1)
Continuous kernel type. One of "gaussian" (default), "epanechnikov", "uniform".

ckerorder :: integer(1)
Continuous Kernel order (only for gaussian and epanechnikov). One of 2 (default), 4, 6, 8.

ukertype :: character(1)
Unordered categorical kernel type. One of "aitchisonaitken" (default) or "liracine".

okertype :: character(1)
Ordered kernel type. One of "liracine" (default), "wangvanryzin".

nmulti :: integer(1)
Number of random restarts for cross validation likelihood optimization. Default: number of features, at most 5. 0 disables restarts.

remin :: logical(1)
Restart from located minima; Default TRUE.

itmax :: integer(1)
Max optimization iters. Default 10000.

ftol :: numeric(1)
Optimization y-value relative tolerance, default 10 * sqrt(.Machine$double.eps).

tol :: numeric(1)
Optimization x-value relative tolerance, default 10000 * sqrt(.Machine$double.eps).

small :: numeric(1)
Optimization x-value absolute tolerance (?), default 1000 * sqrt(.Machine$double.eps).

min_bandwidth :: numeric(1)
Minimum bandwidth (all kerneltypes). Not part of np::npudensbw. Default 0. BOHB has this at 1e-3.

sampling_bw_factor :: numeric(1)
Oversmoothing bandwidth factor for sampling. Not part of np::npudensbw. Default 1. BOHB has this at 3.

TODO: could also implement lb{c,d}.{dir,init}, {c,d}fac.{dir,init}, dfc.dir, hbd.dir, hb{c,d}.init, init{c,d}.dir, scale.init.categorical.sample.

TODO: normal-reference seems to disable optimization, consider putting that as a dependency on the optimization-related params.

TODO: could add dependency for ckerorder on not-uniform.

See also

Super classes

mlr3::Learner -> smashy::LearnerDensity -> LearnerDensityNP

Methods

Public methods

LearnerDensityNP$new()

LearnerDensityNP$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

smashy::LearnerDensity$sample()

Method new()

Initialize the LearnerDensityNP object.

Usage

LearnerDensityNP$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerDensityNP$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.