Split Regression Task into two Density Tasks

Create two TaskDensity from a TaskRegr: one comprising the upper alpha fraction, the other the 1 - alpha lower fraction (but at least min_size).

Format

R6Class inheriting from PipeOp.

Construction

PipeOpDensitySplit$new(id = "densitysplit", param_vals = list())

id :: character(1)
Identifier of the resulting object, default "densityratio".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

PipeOpDensitySplit has one input channels named "input", taking a TaskRegr during both training and prediction.

PipeOpDensitySplit has two output channels, "top" and "bottom", both TaskDensity during both training and prediction.

State

The $state is left empty (list()).

Parameters

The parameters are the parameters inherited from the PipeOp, as well as:

alpha :: numeric(1)
What proportion of values to consider 'good'. BOHB has this at 0.15.

min_size :: integer(1)
Minimum size of both TaskDensity to create. Initialized to 1.

Internals

Can be used with PipeOpDensityRatio.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

See also

Super class

mlr3pipelines::PipeOp -> PipeOpDensitySplit

Methods

Public methods

PipeOpDensitySplit$new()

PipeOpDensitySplit$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Usage

PipeOpDensitySplit$new(id = "densitysplit", param_vals = list())

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpDensitySplit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.