Stratify Training by Feature

Subset the input as a Multiplicity, causing subsequent PipeOps to be executed independently for each stratum according to stratify_feature.

# TODO: missing: 'build only the model with the highest stratify param value'

Format

R6Class object inheriting from PipeOp.

Construction

PipeOpStratify$new(id = "stratify", param_vals = list())

id :: character(1) Identifier of the resulting object, default "stratify".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

PipeOpStratify has one input channel named "input", taking a Task, both during training and prediction.

PipeOpStratify has two output channels; one named "output", returning Task subsets as a Multiplicity during training and prediction; and one named "fallback", containing the whole Task during training and the subset of the input Task that doesn't fit any stratum during prediction.

State

The $state is a named list with the following entries:

stratify_values :: any
Vector indicating the values of features by which

Parameters

stratify_feature :: character(1)
Name of the feature to stratify by.

min_size :: integer(1)
Minimum number of samples for a stratification level so that a model gets trained. Initializedto 1.

predict_choice :: character(1)
How to choose the model with which to make predictions. "exact": only model with exactly matching stratification value used. "exact_or_greater": if exact model was not trained, take the model with next higher stratify value. "exact_or_less": if exact model was not trained, take the model with the next lower stratify value. "nearest": Take the model with nearest stratify value, whether lower or higher. Initialized to "exact".
Whenever a model with exact / greater / smaller / any stratify value is not present, the "fallback" model is used.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

See also

Super class

mlr3pipelines::PipeOp -> PipeOpStratify

Methods

Public methods

PipeOpStratify$new()

PipeOpStratify$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Usage

PipeOpStratify$new(id = "stratify", param_vals = list())

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpStratify$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.