Emulate the BOHB optimizer.

Create a OptimizerSmashy (or TunerSmashy) object which behaves as the BOHB optimizer presented by Falkner et al. (2018).

smashy_as_bohb(search_space, eta = 3, rho = 1/3, ns = 64, type = "Optimizer")

Arguments

search_space

(ParamSet)
search space for which OptimizerSmashy should be configured, has one parameter tagged 'budget'. Must be the same search space as of the OptimInstanceSingleCrit being optimized with the resulting optimizer.
The budget-parameter must be in log-scale; in most cases this means that budget_to_logscale() must be called on the search space. Note that this should be done before OptimInstanceSingleCrit is created with it, see examples.

eta

(numeric(1))
Eta-parameter of BOHB: Factor of budget increase and at the same time, one over fraction of configurations that survive, for each batch evaluation. Default 3.

rho

(numeric(1))
Rho-parameter of BOHB: fraction of configurations sampled randomly without the aid of the surrogate model.
Note that this implementation differs from the implementation by Falkner et al. (2018), since the kernel density sampler of BOHB is used even for the randomly interleaved configurations here, whereas Falkner et al. (2018) samples these points uniformly at random.

ns

(integer(1))
Surrogate random search rate: How many randomly sampled configurations to evaluate on the surrogate model to choose a single configuration to evaluate on the true objective function.

type

(character(1))
One of "Optimizer" or "Tuner". What class to return: An Optimizer for optimizing OptimInstanceSingleCrit objects, or a Tuner for tuning mlr3::Learners. Defaults to "Optimizer".

Value

OptimizerSmashy or TunerSmashy, depending on the type argument: The configured Smashy optimizer.

See also

Examples

# Define the objective to optimize
# The 'budget' here simulates averaging 'b' samples from a noisy function
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    z <- z + rnorm(1, sd = 1 / sqrt(xs$b))
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4), b = p_int(1)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

search_space_proto = objective$domain$search_space(list(
  x = to_tune(),
  y = to_tune(),
  b = to_tune(p_int(1, 2^10, tags = "budget"))
))

search_space = budget_to_logscale(search_space_proto)

# Get a new OptimInstance. Here we determine that the optimizatoin goes
# for 10 full budget evaluations (10 * 2^10)
oi <- OptimInstanceSingleCrit$new(objective,
  search_space = search_space,
  terminator = trm("budget", budget = 10 * 2^10)
)

bohb = smashy_as_bohb(search_space)

bohb$optimize(oi)
#>            x         y         b  x_domain      Obj
#> 1: -1.049624 -1.874318 0.3397981 <list[3]> 3.649478