Emulate the Hyperband optimizer.

Create a OptimizerSmashy (or TunerSmashy) object which behaves as the Hyperband optimizer presented by Li et al. (2018).

smashy_as_hyperband(search_space, eta = 3, type = "Optimizer")

Arguments

search_space

(ParamSet)
search space for which OptimizerSmashy should be configured, has one parameter tagged 'budget'. Must be the same search space as of the OptimInstanceSingleCrit being optimized with the resulting optimizer.
The budget-parameter must be in log-scale; in most cases this means that budget_to_logscale() must be called on the search space. Note that this should be done before OptimInstanceSingleCrit is created with it, see examples.

eta

(numeric(1))
Eta-parameter of Hyperband: Factor of budget increase and at the same time, one over fraction of configurations that survive, for each batch evaluation. Default 3.

type

(character(1))
One of "Optimizer" or "Tuner". What class to return: An Optimizer for optimizing OptimInstanceSingleCrit objects, or a Tuner for tuning mlr3::Learners. Defaults to "Optimizer".

Value

OptimizerSmashy or TunerSmashy, depending on the type argument: The configured Smashy optimizer.

See also

Examples

# Define the objective to optimize
# The 'budget' here simulates averaging 'b' samples from a noisy function
objective <- ObjectiveRFun$new(
  fun = function(xs) {
    z <- exp(-xs$x^2 - xs$y^2) + 2 * exp(-(2 - xs$x)^2 - (2 - xs$y)^2)
    z <- z + rnorm(1, sd = 1 / sqrt(xs$b))
    list(Obj = z)
  },
  domain = ps(x = p_dbl(-2, 4), y = p_dbl(-2, 4), b = p_int(1)),
  codomain = ps(Obj = p_dbl(tags = "maximize"))
)

search_space_proto = objective$domain$search_space(list(
  x = to_tune(),
  y = to_tune(),
  b = to_tune(p_int(1, 3^3, tags = "budget"))
))

search_space = budget_to_logscale(search_space_proto)

# Get a new OptimInstance. Here we determine that the optimizatoin goes
# for 10 full budget evaluations (10 * 2^100)
oi <- OptimInstanceSingleCrit$new(objective,
  search_space = search_space,
  terminator = trm("gens", generations = 10)
)

hb = smashy_as_hyperband(search_space)

hb$optimize(oi)
#>           x        y b  x_domain      Obj
#> 1: 1.971429 1.786419 0 <list[3]> 2.485885

# Evaluations done by Hyperband:
## 1st bracket:
# 1st  generation: 27 @ budget  1 -> b = log(1)  = 0
# 2nd  generation:  9 @ budget  3 -> b = log(3)  = 1.1
# 3rd  generation:  3 @ budget  9 -> b = log(9)  = 2.2
# 4th  generation:  1 @ budget 27 -> b = log(27) = 3.3
## 2nd bracket:
# 5th  generation: 13 @ budget  3 (b = 1.1)
# 6th  generation:  4 @ budget  9 (b = 2.2)
# 7th  generation:  1 @ budget 27 (b = 3.3)
## 3rd bracket:
# 8th  generation:  6 @ budget  9 (b = 2.2)
# 9th  generation:  2 @ budget 27 (b = 3.3)
## 4th bracket:
# 10th generation:  4 @ budget 27 (b = 3.3)

oi$archive$data[, .N, by = c("b", "dob")]
#>            b dob  N
#>  1: 0.000000   1 27
#>  2: 1.098612   2  9
#>  3: 2.197225   3  3
#>  4: 3.295837   4  1
#>  5: 1.098612   5 13
#>  6: 2.197225   6  4
#>  7: 3.295837   7  1
#>  8: 2.197225   8  6
#>  9: 3.295837   9  2
#> 10: 3.295837  10  4

# full eval log:
oi$archive
#> <Archive>
#>           x      y   b dob eol     Obj              timestamp batch_nr
#>  1:  0.4133  0.247 0.0   1   1  0.4954 2022-08-10 05:18:08.28        1
#>  2: -0.8563 -0.063 0.0   1   1 -0.3748 2022-08-10 05:18:08.28        1
#>  3:  2.6144  1.778 0.0   1   2  1.5660 2022-08-10 05:18:08.28        1
#>  4: -1.6796  2.558 0.0   1   1  0.5124 2022-08-10 05:18:08.28        1
#>  5:  0.9994  2.011 0.0   1   1  0.9126 2022-08-10 05:18:08.28        1
#>  6:  1.4233  2.433 0.0   1   2  1.4975 2022-08-10 05:18:08.28        1
#>  7: -0.1525  3.915 0.0   1   1  0.3195 2022-08-10 05:18:08.28        1
#>  8:  3.4476  3.642 0.0   1   1  1.0233 2022-08-10 05:18:08.28        1
#>  9: -0.6769  0.750 0.0   1   1 -0.5379 2022-08-10 05:18:08.28        1
#> 10: -1.1940  2.806 0.0   1   1  1.0813 2022-08-10 05:18:08.28        1
#> 11:  3.8742  1.394 0.0   1   2  1.6593 2022-08-10 05:18:08.28        1
#> 12: -0.0058 -1.395 0.0   1   2  1.1949 2022-08-10 05:18:08.28        1
#> 13:  1.6162  1.266 0.0   1   2  1.5987 2022-08-10 05:18:08.28        1
#> 14: -0.8337 -0.076 0.0   1   1  0.8005 2022-08-10 05:18:08.28        1
#> 15: -0.1810  3.102 0.0   1   1 -0.0244 2022-08-10 05:18:08.28        1
#> 16:  2.4962  1.087 0.0   1   2  2.2645 2022-08-10 05:18:08.28        1
#> 17:  0.7767 -0.624 0.0   1   1 -0.3622 2022-08-10 05:18:08.28        1
#> 18:  2.0499 -1.610 0.0   1   1  0.5654 2022-08-10 05:18:08.28        1
#> 19:  1.9714  1.786 0.0   1   2  2.4859 2022-08-10 05:18:08.28        1
#> 20:  3.9141  0.580 0.0   1   1 -0.2671 2022-08-10 05:18:08.28        1
#> 21:  3.0608 -0.788 0.0   1   1 -1.0536 2022-08-10 05:18:08.28        1
#> 22: -0.3861 -0.560 0.0   1   2  2.4542 2022-08-10 05:18:08.28        1
#> 23: -0.7599  1.972 0.0   1   1 -0.1804 2022-08-10 05:18:08.28        1
#> 24: -1.4380  1.601 0.0   1   1 -0.6644 2022-08-10 05:18:08.28        1
#> 25: -0.9781  1.613 0.0   1   1 -1.4455 2022-08-10 05:18:08.28        1
#> 26: -1.8966  3.086 0.0   1   2  2.2068 2022-08-10 05:18:08.28        1
#> 27:  0.8713  3.352 0.0   1   1 -0.2206 2022-08-10 05:18:08.28        1
#> 28:  2.6144  1.778 1.1   2   3  1.7313 2022-08-10 05:18:08.33        2
#> 29:  1.4233  2.433 1.1   2   3  0.6840 2022-08-10 05:18:08.33        2
#> 30:  3.8742  1.394 1.1   2   2  0.4090 2022-08-10 05:18:08.33        2
#> 31: -0.0058 -1.395 1.1   2   2 -0.2858 2022-08-10 05:18:08.33        2
#> 32:  1.6162  1.266 1.1   2   2  0.4853 2022-08-10 05:18:08.33        2
#> 33:  2.4962  1.087 1.1   2   2  0.3130 2022-08-10 05:18:08.33        2
#> 34:  1.9714  1.786 1.1   2   3  2.0889 2022-08-10 05:18:08.33        2
#> 35: -0.3861 -0.560 1.1   2   2 -0.3604 2022-08-10 05:18:08.33        2
#> 36: -1.8966  3.086 1.1   2   2 -0.3232 2022-08-10 05:18:08.33        2
#> 37:  2.6144  1.778 2.2   3   3  1.4508 2022-08-10 05:18:08.37        3
#> 38:  1.4233  2.433 2.2   3   3  1.2238 2022-08-10 05:18:08.37        3
#> 39:  1.9714  1.786 2.2   3   4  2.1384 2022-08-10 05:18:08.37        3
#> 40:  1.9714  1.786 3.3   4   4  2.1441 2022-08-10 05:18:08.41        4
#> 41:  2.5631  1.183 1.1   5   5  0.1605 2022-08-10 05:18:08.46        5
#> 42:  3.5839  2.966 1.1   5   6  1.2009 2022-08-10 05:18:08.46        5
#> 43:  2.1577  1.278 1.1   5   6  1.5865 2022-08-10 05:18:08.46        5
#> 44:  0.6228  1.799 1.1   5   5  0.0706 2022-08-10 05:18:08.46        5
#> 45: -1.9861  0.156 1.1   5   5 -0.8921 2022-08-10 05:18:08.46        5
#> 46:  0.5109  0.872 1.1   5   5 -0.0037 2022-08-10 05:18:08.46        5
#> 47:  3.0513  1.967 1.1   5   5  1.1133 2022-08-10 05:18:08.46        5
#> 48:  1.1986  3.499 1.1   5   5  0.5950 2022-08-10 05:18:08.46        5
#> 49: -1.3653  2.982 1.1   5   5  0.3351 2022-08-10 05:18:08.46        5
#> 50:  0.6026  0.510 1.1   5   5  0.3150 2022-08-10 05:18:08.46        5
#> 51: -1.7462  2.150 1.1   5   5 -0.2726 2022-08-10 05:18:08.46        5
#> 52: -0.5275  0.632 1.1   5   6  1.1953 2022-08-10 05:18:08.46        5
#> 53:  2.1318  1.827 1.1   5   6  2.3056 2022-08-10 05:18:08.46        5
#> 54:  3.5839  2.966 2.2   6   6  0.3855 2022-08-10 05:18:08.51        6
#> 55:  2.1577  1.278 2.2   6   7  1.1579 2022-08-10 05:18:08.51        6
#> 56: -0.5275  0.632 2.2   6   6  0.5742 2022-08-10 05:18:08.51        6
#> 57:  2.1318  1.827 2.2   6   6  1.0308 2022-08-10 05:18:08.51        6
#> 58:  2.1577  1.278 3.3   7   7  1.1880 2022-08-10 05:18:08.55        7
#> 59:  2.4334  3.687 2.2   8   8 -0.0408 2022-08-10 05:18:08.59        8
#> 60: -1.0101 -1.976 2.2   8   8  0.0243 2022-08-10 05:18:08.59        8
#> 61:  1.2188  3.780 2.2   8   8 -0.2680 2022-08-10 05:18:08.59        8
#> 62:  2.3130  3.091 2.2   8   9  0.6743 2022-08-10 05:18:08.59        8
#> 63:  1.2493  0.685 2.2   8   9  0.5033 2022-08-10 05:18:08.59        8
#> 64:  1.0534  3.520 2.2   8   8 -0.5551 2022-08-10 05:18:08.59        8
#> 65:  2.3130  3.091 3.3   9   9  0.5460 2022-08-10 05:18:08.64        9
#> 66:  1.2493  0.685 3.3   9   9  0.1779 2022-08-10 05:18:08.64        9
#> 67:  0.3863  2.494 3.3  10  NA  0.1659 2022-08-10 05:18:08.70       10
#> 68:  2.8825  0.365 3.3  10  NA -0.0322 2022-08-10 05:18:08.70       10
#> 69: -0.9977  3.712 3.3  10  NA -0.4909 2022-08-10 05:18:08.70       10
#> 70: -1.3301  0.826 3.3  10  NA  0.0510 2022-08-10 05:18:08.70       10
#>           x      y   b dob eol     Obj              timestamp batch_nr